checkpoint progress in integrating linear algebra decompositions into nCompiler
Open development need is to add better support to the nCompiler nClass system.  The system is incomplete as it does not cleanly create internal nCompiler types for predefined nClass objects, nor adequate support for C++ compiling and linking.  The draft code has some rudimentary, non-portable and likely non-sustainable solutions to the problem, as a reference.